BLD: bluesky and ophyd are required for test fixtures.
We actually need an unmerged branch of ophyd, but we have to handle
that in ``requirements-dev.txt`` in the downstream packages. When
the required update to ophyd is finally released, this will be good to
go and downstream packages can remove the special dependency in their
``requirements-dev.txt`` at their leisure.